1. Barcode scanner with ZXING library (all types of barcodes included)
