+++ .idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="..\:/Users/admin/AndroidStudioProjects/VanillaMVVMProject/app/src/main/res/layout/activity_barcode_scanner.xml" value="0.25416666666666665" />
         <entry key="..\:/Users/admin/AndroidStudioProjects/VanillaMVVMProject/app/src/main/res/layout/activity_main.xml" value="0.24010416666666667" />
+        <entry key="..\:/Users/admin/AndroidStudioProjects/VanillaMVVMProject/app/src/main/res/layout/barcode_layout.xml" value="0.253125" />
         <entry key="..\:/Users/admin/AndroidStudioProjects/VanillaMVVMProject/app/src/main/res/layout/barcode_result.xml" value="0.21428571428571427" />
         <entry key="..\:/Users/admin/AndroidStudioProjects/VanillaMVVMProject/app/src/main/res/layout/barcode_scanner_layout.xml" value="0.253125" />
         <entry key="..\:/Users/admin/AndroidStudioProjects/VanillaMVVMProject/app/src/main/res/layout/content_barcode_scanner.xml" value="0.25416666666666665" />
