+++ app/src/main/java/com/example/vanillamvvmproject/BarcodeScanner.kt
@@ -19,111 +19,19 @@ import com.google.zxing.integration.android.IntentIntegrator
 import com.google.zxing.integration.android.IntentResult
 
 
-/*
+
 class BarcodeScanner : AppCompatActivity()  {
 
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.barcode_scanner_layout)
+        setContentView(R.layout.barcode_layout)
         val  barcodeTextView = findViewById<TextView>(R.id.txt_result)
         val barcodeValue = intent.getStringExtra("barcodeValue")
         barcodeTextView.text = barcodeValue
-
     }
 
-   val btn = findViewById<Button>(R.id.btn1)
-   btn.setOnClickListener(){
-            openCamera()
-            val integrator = IntentIntegrator(this)
-            integrator.setDesiredBarcodeFormats(IntentIntegrator.ALL_CODE_TYPES)
-            integrator.setPrompt("Scan a barcode")
-            integrator.setCameraId(0) // Use the rear camera
-            integrator.setBeepEnabled(false)
-            integrator.setOrientationLocked(false)
-            integrator.setBarcodeImageEnabled(true)
-            integrator.initiateScan()
-
-
-        }
-
-  private val barcodeLauncher: ActivityResultLauncher<Intent> = registerForActivityResult(
-        ActivityResultContracts.StartActivityForResult()
-    ) { result ->
-        val contents: String? =
-            IntentIntegrator.parseActivityResult(result.resultCode, result.data).contents
-        if (contents == null) {
-            Toast.makeText(this, "Cancelled", Toast.LENGTH_LONG).show()
-        } else {
-            Toast.makeText(this, "Scanned: $contents", Toast.LENGTH_LONG).show()
-        }
-    }
-
-
-    fun openCamera(){
-        if (ContextCompat.checkSelfPermission(this@MainActivity,
-                android.Manifest.permission.ACCESS_FINE_LOCATION) !==
-            PackageManager.PERMISSION_GRANTED) {
-            if (ActivityCompat.shouldShowRequestPermissionRationale(this@MainActivity,
-                    android.Manifest.permission.CAMERA)) {
-                ActivityCompat.requestPermissions(this@MainActivity,
-                    arrayOf(android.Manifest.permission.CAMERA), MainActivity.REQUEST_CODE
-                )
-            } else {
-                ActivityCompat.requestPermissions(this@MainActivity,
-                    arrayOf(android.Manifest.permission.CAMERA), MainActivity.REQUEST_CODE
-                )
-            }
-        }
-
-    }
-        override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<out String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        when (REQUEST_CODE) {
-            1 -> {
-                if (grantResults.isNotEmpty() && grantResults[0] ==
-                    PackageManager.PERMISSION_GRANTED) {
-                    if ((ContextCompat.checkSelfPermission(this@MainActivity,
-                            android.Manifest.permission.CAMERA) ===
-                                PackageManager.PERMISSION_GRANTED)) {
-                        Toast.makeText(this, "Permission Granted", Toast.LENGTH_LONG).show()
-                        capturePhoto()
-                        barcodeView.resume()
-                    }
-                } else {
-                    Toast.makeText(this, "Permission Denied", Toast.LENGTH_LONG).show()
-                }
-                return
-            }
-        }
-    }
-
-    fun capturePhoto(){
-        val cameraIntent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
-        startActivityForResult(cameraIntent, REQUEST_CODE)
-    }
-   override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        val result: IntentResult? = IntentIntegrator.parseActivityResult(requestCode, resultCode, data)
-        if (result != null) {
-            val intent = Intent(this, BarcodeScanner::class.java)
-            barcodeLauncher.launch(intent)
-            val barcodeValue = result.contents
-            intent.putExtra("barcodeValue", barcodeValue)
-            startActivity(intent)
-        } else {
-            super.onActivityResult(requestCode, resultCode, data)
-        }
-    }
-
-    companion object{
-        const val REQUEST_CODE = 1
-    }
+}
 
 
 
-
-}*/
